requirements.txt
qiskit
qiskit-ionq
firebase-admin
python-dotenv
matplotlib
